Add List component and update project dependencies
Created a new List component which will be responsible for displaying search results. This component uses Redux's useSelector() to get state from the Redux store. Based on the state, the component either displays a loading message, an error message, or the search results. Changes were also made to .gitignore to exclude Jetbrains files and folders, and updates to Tailwind configuration were added.

Changes the README.md were made as well due to updates to the project dependencies. Several new packages were added and their changes are reflected in the pnpm-lock.yaml file. Some changes in the formatting of strings, from single quotes to double quotes, were put in place in accordance to the teams coding style guide.

The Redux library and the Redux toolkit were added to the list of devDependencies to enable the implementation of Redux in the application. The tailwindcss typography plugin was also added for better styling options in the app's UI components. The project's README file was also updated to reflect these changes and provide updated instructions for getting started with development.