Refactor List component and refactor saga actions
Changes the List component, replacing 'item.id' with 'index' as a key for the component rendering process. This change was made to prevent issues because item IDs can be non-unique in some cases, leading to undesired behavior.

Also, renamed and refactored 'actions.ts' to 'sataActions.ts' to have a better organization and to avoid confusion in our Redux Saga structure since 'actions' is a broad term. This is reflected in 'index.ts' and 'Search.tsx', where the import of a function has been updated to the new file name. Lastly, removed a 'div' wrapper from the input element in Search component, it wasn't necessary and its removal simplifies the code.