Add Material UI icons and clear functionality to Search component
Material UI icons dependency has been added to 'package.json' and 'pnpm-lock.yaml', as to include and utilize them in the application. In 'Search.tsx', Close Icon from Material UI has been incorporated to serve as a clear button for search input. This improves the user interaction and experience as it allows users to easily clear the search input with a single click. The function 'clearHandler' was implemented to clear the input and dispatch the change to the redux store. Lastly, a placeholder text has been added to the search input field for better user guidance.