Remove `deploy` script and simplify `nextConfig`
The `deploy` script was removed from `package.json` because we've changed our deployment strategy. We're no longer deploying to GitHub Pages, so it's unnecessary.

The `nextConfig` in `next.config.js` is also simplified. This change is due to the fact that our basePath is no longer conditionally set on the environment, hence eliminating the need for checks and excessive configuration. This makes the app configuration cleaner and easier to maintain.